% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_cats.R
\name{see_cats}
\alias{see_cats}
\title{See all categories in a blog}
\usage{
see_cats(blog_url = Sys.getenv("BLOG_URL"))
}
\arguments{
\item{blog_url}{}
}
\value{
character vector with available categories on blog home page
}
\description{
Set an R environment variable BLOG_URL to your blog's URL including the https:// if you don't want to enter the URL manually.
}
\examples{
see_cats("https://themockup.blog/")
}
