% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpost.R
\name{newpost}
\alias{newpost}
\title{Create new folder and index.qmd in your blog's posts directory}
\usage{
newpost(headline, root_dir = here::here())
}
\arguments{
\item{root_dir}{string your blog's root directory, defaults to here::here()}

\item{character}{string with planned post headline}
}
\value{
Creates a new folder in posts directory with yyy--mm-dd-headline-in-lowercase where yyyy-mm-dd is current system date
}
\description{
Create new folder and index.qmd in your blog's posts directory
}
